# docs/api/solr-openapi.json
{"openapi":"3.0.3","info":{"title":"Solr Administration API","description":"API endpoints for configuring and managing Solr databases and collections","contact":{"name":"HistText Support","email":"support@histtext.example.com"},"license":{"name":""},"version":"1.0.0"},"paths":{"/api/solr_database_infos":{"get":{"tags":["SolrDatabaseInfo"],"summary":"Retrieves all Solr collection metadata records","description":"Returns configuration information about Solr collections, including\ndescriptions, embedding paths, language settings, and display options.\n\n# Arguments\n* `db` - Database connection pool\n* `config` - Application configuration\n\n# Returns\nHTTP response with all metadata records as JSON","operationId":"get_solr_database_infos","responses":{"200":{"description":"List of SolrDatabaseInfo records with collection metadata","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SolrDatabaseInfo"}}}}},"500":{"description":"Database connection error or query execution failure"}},"security":[{"bearer_auth":[]}]},"post":{"tags":["SolrDatabaseInfo"],"summary":"Creates a new Solr collection metadata record","description":"Validates and stores metadata for a Solr collection, including\ndescription, embedding path, language settings, and display options.\n\n# Arguments\n* `db` - Database connection pool\n* `item` - JSON payload with new metadata\n* `config` - Application configuration\n\n# Returns\nHTTP response with the created record or validation error","operationId":"create_solr_database_info","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewSolrDatabaseInfo"}}},"required":true},"responses":{"201":{"description":"Collection metadata record created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolrDatabaseInfo"}}}},"400":{"description":"Validation error: invalid solr_database_id, collection name too long, or missing required fields"},"500":{"description":"Database connection error, constraint violation, or execution failure"}},"security":[{"bearer_auth":[]}]}},"/api/solr_database_infos/{solr_database_id}/{collection_name}":{"get":{"tags":["SolrDatabaseInfo"],"summary":"Retrieves a specific Solr collection metadata record","description":"Looks up collection metadata by the composite key of\nsolr_database_id and collection_name.\n\n# Arguments\n* `db` - Database connection pool\n* `path` - Path parameters containing database ID and collection name\n* `config` - Application configuration\n\n# Returns\nHTTP response with the matching metadata record or not-found error","operationId":"get_solr_database_info","parameters":[{"name":"solr_database_id","in":"path","required":true,"schema":{"type":"integer","format":"int32"},"example":1},{"name":"collection_name","in":"path","required":true,"schema":{"type":"string"},"example":"users"}],"responses":{"200":{"description":"SolrDatabaseInfo metadata record found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolrDatabaseInfo"}}}},"404":{"description":"No metadata found for the specified database ID and collection"},"500":{"description":"Database connection error or query execution failure"}},"security":[{"bearer_auth":[]}]},"put":{"tags":["SolrDatabaseInfo"],"summary":"Updates an existing Solr collection metadata record","description":"Allows partial updates to collection metadata with\nvalidation of the provided fields.\n\n# Arguments\n* `db` - Database connection pool\n* `path` - Path parameters containing database ID and collection name\n* `item` - JSON payload with fields to update\n* `config` - Application configuration\n\n# Returns\nHTTP response with the updated record or appropriate error","operationId":"update_solr_database_info","parameters":[{"name":"solr_database_id","in":"path","required":true,"schema":{"type":"integer","format":"int32"},"example":1},{"name":"collection_name","in":"path","required":true,"schema":{"type":"string"},"example":"users"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSolrDatabaseInfo"}}},"required":true},"responses":{"200":{"description":"Collection metadata updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolrDatabaseInfo"}}}},"400":{"description":"Validation error: invalid field lengths or formats"},"404":{"description":"No metadata found for the specified database ID and collection"},"500":{"description":"Database connection error or execution failure"}},"security":[{"bearer_auth":[]}]},"delete":{"tags":["SolrDatabaseInfo"],"summary":"Deletes a Solr collection metadata record","description":"Removes metadata for a specific collection if it exists.\n\n# Arguments\n* `db` - Database connection pool\n* `path` - Path parameters containing database ID and collection name\n* `config` - Application configuration\n\n# Returns\nHTTP response with success message or not-found error","operationId":"delete_solr_database_info","parameters":[{"name":"solr_database_id","in":"path","required":true,"schema":{"type":"integer","format":"int32"},"example":1},{"name":"collection_name","in":"path","required":true,"schema":{"type":"string"},"example":"users"}],"responses":{"200":{"description":"Collection metadata record deleted successfully"},"404":{"description":"No metadata found for the specified database ID and collection"},"500":{"description":"Database connection error or execution failure"}},"security":[{"bearer_auth":[]}]}},"/api/solr_database_permissions":{"get":{"tags":["SolrDatabasePermissions"],"summary":"Retrieves all Solr collection permission requirements","description":"Lists all permission requirements defining which permissions\nare needed to access specific Solr collections. This endpoint\nis useful for security auditing and permission management.\n\n# Arguments\n* `db` - Database connection pool\n* `config` - Application configuration\n\n# Returns\nHTTP response with all permission records as JSON","operationId":"get_solr_database_permissions","responses":{"200":{"description":"List of all Solr collection permission requirements","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SolrDatabasePermission"}}}}},"500":{"description":"Database connection error or query execution failure"}},"security":[{"bearer_auth":[]}]},"post":{"tags":["SolrDatabasePermissions"],"summary":"Creates a new Solr collection permission requirement","description":"Defines a new permission requirement for accessing a specific\nSolr collection. This creates an access control rule requiring\nusers to have the specified permission to access the collection.\n\n# Arguments\n* `db` - Database connection pool\n* `item` - JSON payload with the new permission requirement\n* `config` - Application configuration\n\n# Returns\nHTTP response with the created permission record or validation error","operationId":"create_solr_database_permission","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewSolrDatabasePermission"}}},"required":true},"responses":{"201":{"description":"Solr collection permission requirement created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolrDatabasePermission"}}}},"400":{"description":"Validation error: invalid solr_database_id, collection name too long, or permission format invalid"},"500":{"description":"Database connection error, constraint violation, or execution failure"}},"security":[{"bearer_auth":[]}]}},"/api/solr_database_permissions/{solr_database_id}/{collection_name}/{permission}":{"get":{"tags":["SolrDatabasePermissions"],"summary":"Retrieves a specific Solr collection permission requirement","description":"Looks up a permission requirement by its composite key of\nsolr_database_id, collection_name, and permission.\n\n# Arguments\n* `db` - Database connection pool\n* `path` - Path parameters containing the composite key components\n* `config` - Application configuration\n\n# Returns\nHTTP response with the matching permission record or not-found error","operationId":"get_solr_database_permission","parameters":[{"name":"solr_database_id","in":"path","required":true,"schema":{"type":"integer","format":"int32"},"example":1},{"name":"collection_name","in":"path","required":true,"schema":{"type":"string"},"example":"users"},{"name":"permission","in":"path","required":true,"schema":{"type":"string"},"example":"read"}],"responses":{"200":{"description":"Solr collection permission record found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolrDatabasePermission"}}}},"404":{"description":"No permission record found with the specified composite key"},"500":{"description":"Database connection error or query execution failure"}},"security":[{"bearer_auth":[]}]},"delete":{"tags":["SolrDatabasePermissions"],"summary":"Deletes a Solr collection permission requirement","description":"Removes a permission requirement for accessing a specific collection,\neffectively making that collection accessible without the specified\npermission (unless other permission requirements exist).\n\n# Arguments\n* `db` - Database connection pool\n* `path` - Path parameters containing the composite key components\n* `config` - Application configuration\n\n# Returns\nHTTP response with success message or not-found error","operationId":"delete_solr_database_permission","parameters":[{"name":"solr_database_id","in":"path","required":true,"schema":{"type":"integer","format":"int32"},"example":1},{"name":"collection_name","in":"path","required":true,"schema":{"type":"string"},"example":"users"},{"name":"permission","in":"path","required":true,"schema":{"type":"string"},"example":"read"}],"responses":{"200":{"description":"Solr collection permission requirement deleted successfully"},"404":{"description":"No permission record found with the specified composite key"},"500":{"description":"Database connection error or execution failure"}},"security":[{"bearer_auth":[]}]}},"/api/solr_databases":{"get":{"tags":["SolrDatabases"],"summary":"Retrieves all Solr database configurations","description":"Lists all Solr instance configurations, including server details\nand port mappings, for administration and display purposes.\n\n# Arguments\n* `db` - Database connection pool\n* `config` - Application configuration\n\n# Returns\nHTTP response with all configurations as JSON","operationId":"get_solr_databases","responses":{"200":{"description":"List of all Solr database configurations with connection details","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SolrDatabase"}}}}},"500":{"description":"Database connection error or query execution failure"}},"security":[{"bearer_auth":[]}]},"post":{"tags":["SolrDatabases"],"summary":"Creates a new Solr database configuration","description":"Stores connection details for a Solr instance, including URL,\nserver port (where Solr is running), and local port (for SSH tunneling).\n\n# Arguments\n* `db` - Database connection pool\n* `item` - JSON payload with new configuration details\n* `config` - Application configuration\n\n# Returns\nHTTP response with the created configuration or validation error","operationId":"create_solr_database","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewSolrDatabase"}}},"required":true},"responses":{"201":{"description":"Solr database configuration created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolrDatabase"}}}},"400":{"description":"Bad request, validation failed: invalid name length, empty URL, or port numbers outside valid range (1-65535)"},"500":{"description":"Internal server error: database connection failure or constraint violation"}},"security":[{"bearer_auth":[]}]}},"/api/solr_databases/{id}":{"get":{"tags":["SolrDatabases"],"summary":"Retrieves a specific Solr database configuration","description":"Gets detailed connection information for a single Solr instance\nidentified by its unique ID.\n\n# Arguments\n* `db` - Database connection pool\n* `path` - Path parameter containing the database ID\n* `config` - Application configuration\n\n# Returns\nHTTP response with the requested configuration or not-found error","operationId":"get_solr_database_by_id","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"200":{"description":"Solr database configuration found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolrDatabase"}}}},"404":{"description":"No Solr database configuration found with the specified ID"},"500":{"description":"Database connection error or query execution failure"}},"security":[{"bearer_auth":[]}]},"put":{"tags":["SolrDatabases"],"summary":"Updates an existing Solr database configuration","description":"Modifies connection details for a Solr instance. All fields\nare optional to support partial updates.\n\n# Arguments\n* `db` - Database connection pool\n* `path` - Path parameter containing the database ID\n* `item` - JSON payload with fields to update\n* `config` - Application configuration\n\n# Returns\nHTTP response with the updated configuration or appropriate error","operationId":"update_solr_database","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSolrDatabase"}}},"required":true},"responses":{"200":{"description":"Solr database configuration updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolrDatabase"}}}},"400":{"description":"Bad request, validation failed: invalid name length, empty URL, or invalid port range"},"404":{"description":"Solr database configuration not found"},"500":{"description":"Internal server error: database connection failure or constraint violation"}},"security":[{"bearer_auth":[]}]},"delete":{"tags":["SolrDatabases"],"summary":"Deletes a Solr database configuration","description":"Removes connection details for a Solr instance identified by its ID.\n\n# Arguments\n* `db` - Database connection pool\n* `path` - Path parameter containing the database ID\n* `config` - Application configuration\n\n# Returns\nHTTP response with success message or not-found error","operationId":"delete_solr_database","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"200":{"description":"Solr database configuration deleted successfully"},"404":{"description":"Solr database configuration not found"},"500":{"description":"Internal server error: database connection failure or constraint violation"}},"security":[{"bearer_auth":[]}]}},"/api/solr_databases/{id}/connect_ssh":{"post":{"tags":["Solr Database Management"],"summary":"API handler for connecting a specific Solr database via SSH","description":"Exposes an endpoint that requires admin permission to establish\nan SSH connection to a specific Solr database.\n\n# Arguments\n* `path` - Path parameter containing the Solr database ID\n* `pool` - Database connection pool\n* `app_state` - Application state\n\n# Returns\nHTTP response indicating success or failure","operationId":"connect_solr_database_ssh","parameters":[{"name":"id","in":"path","description":"Solr database ID to connect to","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"SSH connection established successfully"},"404":{"description":"Solr database not found"},"500":{"description":"Failed to establish SSH connection"}},"security":[{"Bearer":[]}]}},"/api/stats":{"get":{"tags":["Stats"],"summary":"Retrieves application dashboard statistics","description":"Aggregates metrics from multiple tables to provide high-level\ninformation about users, roles, permissions, and system configuration.\nRequires admin permissions to access.\n\n# Arguments\n* `db` - Database connection pool\n* `config` - Application configuration\n\n# Returns\nHTTP response with dashboard metrics as JSON","operationId":"get_dashboard_stats","responses":{"200":{"description":"Dashboard statistics with user, role, permission, and system counts","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardStats"}}}},"500":{"description":"Database connection error or query execution failure"}},"security":[{"bearer_auth":[]}]}}},"components":{"schemas":{"DashboardStats":{"type":"object","description":"Dashboard statistics model\n\nRepresents high-level metrics about the application,\nincluding user counts, role counts, and system statistics.","required":["active_users","total_users","total_roles","total_permissions","solr_collections","embedding_files_loaded","embedding_collections_cached","embedding_words_loaded"],"properties":{"active_users":{"type":"integer","format":"int64","description":"Number of activated user accounts","example":42},"embedding_collections_cached":{"type":"integer","description":"Number of collections with cached embeddings","example":5,"minimum":0},"embedding_files_loaded":{"type":"integer","description":"Number of embedding files currently loaded in memory","example":2,"minimum":0},"embedding_words_loaded":{"type":"integer","description":"Total number of words loaded in embedding cache","example":400000,"minimum":0},"solr_collections":{"type":"integer","format":"int64","description":"Number of Solr collections configured","example":3},"total_permissions":{"type":"integer","format":"int64","description":"Number of distinct permissions in the system","example":10},"total_roles":{"type":"integer","format":"int64","description":"Number of distinct roles in the system","example":5},"total_users":{"type":"integer","format":"int64","description":"Total number of user accounts (active and inactive)","example":100}}},"NewSolrDatabase":{"type":"object","description":"Data for creating a new Solr database configuration","required":["name","url","server_port","local_port"],"properties":{"local_port":{"type":"integer","format":"int32","description":"Local port to map the SSH tunnel","example":7574},"name":{"type":"string","description":"Human-readable name for this Solr instance","example":"primary-solr"},"server_port":{"type":"integer","format":"int32","description":"Remote port where Solr is running","example":8983},"url":{"type":"string","description":"SSH connection URL or hostname","example":"http://localhost:8983/solr"}}},"NewSolrDatabaseInfo":{"type":"object","description":"Data for creating a new Solr collection metadata record","required":["solr_database_id","collection_name","description","embeddings"],"properties":{"collection_name":{"type":"string","description":"Name of the Solr collection","example":"users"},"description":{"type":"string","description":"Human-readable description of the collection","example":"User profiles index"},"embeddings":{"type":"string","description":"Path to word embedding file or \"default\"/\"none\"","example":"/data/embeddings/users.vec"},"lang":{"type":"string","description":"Language code for the collection","example":"en","nullable":true},"solr_database_id":{"type":"integer","format":"int32","description":"ID of the parent Solr database","example":1},"text_field":{"type":"string","description":"Primary text field for searching and analysis","example":"text_content","nullable":true},"to_not_display":{"type":"array","items":{"type":"string","nullable":true},"description":"Fields to exclude from display in UI","example":["internal_notes",null],"nullable":true},"tokenizer":{"type":"string","description":"Tokenizer to use","example":"standard","nullable":true}}},"NewSolrDatabasePermission":{"type":"object","description":"Data for creating a new Solr database permission","required":["solr_database_id","collection_name","permission"],"properties":{"collection_name":{"type":"string","description":"Name of the Solr collection","example":"users"},"permission":{"type":"string","description":"Permission required to access this collection","example":"read"},"solr_database_id":{"type":"integer","format":"int32","description":"ID of the parent Solr database","example":1}}},"SolrDatabase":{"type":"object","description":"Solr database configuration record\n\nContains connection details for a Solr instance, including\nURL, port mappings for SSH tunneling, and metadata.","required":["id","name","url","server_port","local_port","created_at","updated_at"],"properties":{"created_at":{"type":"string","format":"date-time","description":"When this configuration was created","example":"2025-01-15T12:34:56"},"id":{"type":"integer","format":"int32","description":"Unique identifier","example":1},"local_port":{"type":"integer","format":"int32","description":"Local port to map the SSH tunnel","example":7574},"name":{"type":"string","description":"Human-readable name for this Solr instance","example":"primary-solr"},"server_port":{"type":"integer","format":"int32","description":"Remote port where Solr is running","example":8983},"updated_at":{"type":"string","format":"date-time","description":"When this configuration was last updated","example":"2025-02-20T08:22:10"},"url":{"type":"string","description":"SSH connection URL or hostname","example":"http://localhost:8983/solr"}}},"SolrDatabaseInfo":{"type":"object","description":"Solr collection metadata record\n\nContains configuration and display information for a Solr collection,\nidentified by the combination of solr_database_id and collection_name.","required":["solr_database_id","collection_name","description","embeddings"],"properties":{"collection_name":{"type":"string","description":"Name of the Solr collection","example":"users"},"description":{"type":"string","description":"Human-readable description of the collection","example":"User profiles index"},"embeddings":{"type":"string","description":"Path to word embedding file or \"default\"/\"none\"","example":"/data/embeddings/users.vec"},"lang":{"type":"string","description":"Language code for the collection (e.g. \"en\", \"fr\")","example":"en","nullable":true},"solr_database_id":{"type":"integer","format":"int32","description":"ID of the parent Solr database","example":1},"text_field":{"type":"string","description":"Primary text field for searching and analysis","example":"text_content","nullable":true},"to_not_display":{"type":"array","items":{"type":"string","nullable":true},"description":"Fields to exclude from display in UI","example":["internal_notes",null],"nullable":true},"tokenizer":{"type":"string","description":"Tokenizer to use (e.g. \"standard\", \"whitespace\")","example":"standard","nullable":true}}},"SolrDatabasePermission":{"type":"object","description":"Solr database permission record\n\nRepresents a permission requirement for accessing a specific Solr collection.\nEach record creates an access control rule requiring users to have the specified\npermission to access the collection.","required":["solr_database_id","collection_name","permission"],"properties":{"collection_name":{"type":"string","description":"Name of the Solr collection","example":"users"},"permission":{"type":"string","description":"Permission required to access this collection","example":"read"},"solr_database_id":{"type":"integer","format":"int32","description":"ID of the parent Solr database","example":1}}},"UpdateSolrDatabase":{"type":"object","description":"Data for updating an existing Solr database configuration\n\nAll fields are optional to support partial updates.","properties":{"local_port":{"type":"integer","format":"int32","description":"Updated local port","example":8983,"nullable":true},"name":{"type":"string","description":"Updated name","example":"primary-solr-renamed","nullable":true},"server_port":{"type":"integer","format":"int32","description":"Updated server port","example":7574,"nullable":true},"url":{"type":"string","description":"Updated URL","example":"http://localhost:7574/solr","nullable":true}}},"UpdateSolrDatabaseInfo":{"type":"object","description":"Data for updating an existing Solr collection metadata record\n\nAll fields are optional to support partial updates.","properties":{"description":{"type":"string","description":"Updated description","example":"Updated description","nullable":true},"embeddings":{"type":"string","description":"Updated embedding path","example":"/data/embeddings/updated.vec","nullable":true},"lang":{"type":"string","description":"Updated language code","example":"fr","nullable":true},"text_field":{"type":"string","description":"Updated text field","example":"content_field","nullable":true},"to_not_display":{"type":"array","items":{"type":"string","nullable":true},"description":"Updated list of fields to exclude from display","example":["secret",null],"nullable":true},"tokenizer":{"type":"string","description":"Updated tokenizer setting","example":"whitespace","nullable":true}}}},"securitySchemes":{"bearer_auth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"security":[{"bearer_auth":[]}],"tags":[{"name":"SolrDatabases","description":"Solr instance connection configuration and management"},{"name":"SolrDatabaseInfo","description":"Collection metadata and configuration"},{"name":"SolrDatabasePermissions","description":"Collection-level access control"},{"name":"Stats","description":"System-wide statistics and monitoring"}]}